Added 24-bit ansi support
If a 24-bit RGB value is embedded in a text string, ex [red.greeen.blue], but we are currently in an 8-bit mode the 24-bit RGB value will automatically be convert to the closest match in the 8-bit palette. If in a 4-bit mode it will be converted to the nearest standard color. So you can always specify 24-bit RGB colors and rely on automatic conversion if necessary.

The ansi24 code also includes a wrapper for C standard library printf formatting. Only string, integer and double types are supported. Note that integer types include characters, %c. If a plain %c is used without any formatting the wrapper will allow Unicode codepoints to be specified, this is backwards compatible with ascii. If any formating is specified Unicode support will not be attempted.